fix #1: draw exit arrows outside narrow floor
In 8-cell-tall maps, the exit is all cells off the top of the floor.
But there's no visual indication of where this exit is, as nocash
and Myrrh-stache pointed out to me.  So I allocated tile $6E in
CHR RAM to hold arrows that appear at the top left and right outside
the floor once the exit opens.