quieter roll sound; replace land sound
I made the rolling sound an octave lower and reduced the volume to
minimum.  If I add music, rolling sound will not be played.

After looking for inspiration from real-world rolling animals, such
as the armadillo, pangolin, and hedgehog, I decided that she's not
as rigid as a boulder per se, even if there is some callus in her
kind's skin from rolling all the time.  This changes what landing
needs to sound like. I'm still not sure whether I need a sound when
rolling over the border from one cell to the next.

On pre-GBA systems, a pulse channel turning off has a fairly
noticeable click, which is louder if the channel is quieter when
turning off.  Is the DAC going to 7.5 or something?