calculate odds of seeing no dead ends
Earning the "Sink it" achievement (see #7) requires seeing a
dead end.  ISSOtm raised the possibility that the player might not
see a single dead end over the course of a game.  To investigate
the likelihood of this, I added some calculations to the Python
prototype of the floor generator and solver:

- Add distinction between dead end and unreachable
- Count OK floors of each size with at least one dead end
- Calculate odds of seeing one or more dead ends

Result from one run:

- 2x8: 17%
- 4x4: 26%
- 6x4: 40%
- 6x6: 52%
- 8x6: 58%
- Chance of seeing at least one dead end: 92%