earned achievement slides in over status bar
Each achievement is added to a queue.  The name of the achievement
at the front of the queue is drawn to a reserved area of VRAM and
then displayed in a window that slides a couple pixels past the
status bar to break up players' box blindness.

The only part of #7 left to build is "No scope".