grab tiles from elsewhere in a sprite sheet
Allow a cel to read a strip from elsewhere on a sprite sheet and move
the strip into the cel.  Strips read from elsewhere are not subject
to the cel's clipping rectangle, and in a cel with no clipping
rectangle, the destination position is used to guess the hotspot.

The functionality has been tested but is not in use in this game.
Libbet.ec could use this in spots if I were willing to make certain
scanlines 3 sprites wide.