add 90 degree rotation code
I plan to have a much bigger version of Libbet roll across the DMG
Nintendo logo, replacing it with a copyright notice and knocking
away the (R) symbol.  On GBC, the logo just won't be there.

To do this, we store cels for 0 and 45 degree rotations, rotate them
in software to 90 and 135 degrees, and HV-flip the sprites to 180,
215, 270, and 315 degrees.  The rotation code appears to work (but
isn't called yet).