roll and wipe away console maker logo
The game starts with the console maker logo displayed as a 96x16
pixel image in the center of the screen.  (On GBC and GBA, this
area is blank for all games except "X".)  Add animation of Libbet
rolling past the logo and wiping it away.

Uses an hblank memcpy that moves 8 bytes per scanline from a
decompression buffer to VRAM, or up to 1K per frame.

Future directions:
- Add copyright notice
- Add knocking away trademark symbol