change black cell texture to a grid
The cross on the black cell wasn't quite well received.  Others
on gbdev thought it didn't fit in with the other patterns.
First I tried removing the center row from each group of 6 pixels
in the cross, turning it into 2 ridges like on the DMG's Control
Pad.  Then I realized the real problem: the cross is a symbol, not
a texture.  So I separated out the pairs of pixels in each ridge,
and it started to look like a grid.  Hence this.