dark gray tile tweaks
Testing on original Game Boy (DMG) revealed that the dark gray
tile needed a bit more definition at the corners to distinguish
it from the black tile.

Also update the 24x24 tileset to reflect new direction for the
16x16 tile and to reflect the expected use of 8px Libbet in the
SMS port, just as most of Martin's ports use an 8px boulder on
a 24x24 cell grid.