extractcels.py: pack frames for streaming
A streaming cel is prefixed with the tile numbers it uses, and then
tile numbers in that cel's strips are remapped to indices into that
array (0 to N-1).  This corresponds to copying tiles 0 to N-1
to VRAM and then drawing the metasprite from the copied tiles.

Streaming tile sheets can have 256 tiles.  Assuming 8x16-pixel
objects, this is the size of Donny from Haunted: Halloween '86.

Another GB project uses streaming.  I'm merging changes related to
streaming from that project in order to keep this repo the primary
source for extractcels.py.